LayoutTests/platform/ios-sim-deprecated/editing/secure-input/removed-password-input-expected.txt

Verify that removing a password input from DOM tree disables secure input state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


A password input is focused:
FAIL testRunner.secureEventInputIsEnabled should be true. Was false.

After deleting the input:
PASS testRunner.secureEventInputIsEnabled is false
PASS successfullyParsed is true

TEST COMPLETE
